Weapon Prefabs functioning
Megaphone works, as do nail gun and BBBat, all that's left is to hook them up to the player controller script. The annoying errors that popped up whenever you clicked on my test player are also gone (it just wanted my shit to be private instead of public), as well as the megaphone projectiles expanding over time and dealing damage based on size.